jess_local_connector/requirements.txt
google-cloud-pubsub==2.18.4